Mall Customer Segmentation
Abstract
A lot of customers buy products from the mall and to generate more revenue for the mall, the
authorities need to attract these customers and for this large amount of capital is required. After
the advertisement, the output is only around 30-40%. Hence customer segmentation comes into
the picture.
Customer Segmentation is a popular application of unsupervised learning and by using this
technique we'll only focus on the potential customers (customers whose probability of buying the
product is very high). With this technique, the output will drastically increase to 90-95%.
Our project aims to build clusters of customers based on their Spending Score and Annual
Income. The algorithm used in this project is K-means.